+++ components/Text.tsx
@@ -3,9 +3,10 @@ import setup from "../src/setup/Text.ts";
 
 /**
  * @todo Styles for prose, and better y-margins.
+ * @todo Change size classes from tailwind's in favor of custom sizes.
  */
 export default function (props: Partial<iText>) {
-  const { c, children, inheritColor, ...p } = setup(props);
+  const { c, type, children, inheritColor, ...p } = setup(props);
 
   return (
     <span {...p} class={c.span}>