+++ src/setup/Footer.ts
@@ -10,7 +10,7 @@ export default (props: Partial<iFooter>) => {
   const p = applyDefaults<iFooter>(defaults, props);
 
   const classes = {
-    footer: cn('pt-4 pb-4 mt-10 border-t-1 border-hierro'),
+    footer: cn('pt-4 pb-4 mt-10 border-t-1 clr-border'),
   };
 
   return { c: classes, ...p };