+++ src/setup/Navigation.ts
@@ -10,7 +10,7 @@ export default (props: Partial<iNavigation>) => {
 
   const classes = {
     wrapper: cn('h-10 relative z-10'),
-    nav: cn('py-2 w-full bg-white fixed border-b-1 border-hierro'),
+    nav: cn('py-2 w-full clr-bg-primary fixed border-b-1 clr-border'),
     title: cn('flex justify-between'),
   };