+++ src/setup/Page.ts
@@ -4,12 +4,14 @@ import { iPage } from '../types/props.ts';
 export default (props: Partial<iPage>) => {
   const defaults: iPage = {
     children: undefined,
+    theme: 'newspaper',
+    darkMode: false,
   };
 
   const p = applyDefaults<iPage>(defaults, props);
 
   const classes = {
-    wrapper: cn('page font-mono'),
+    wrapper: cn('page font-mono', p.darkMode ? 'dark' : 'light', p.theme),
   };
 
   return { c: classes, ...p };