+++ src/setup/TextArea.ts
@@ -16,10 +16,10 @@ export default (props: Partial<iTextArea>) => {
       'outline-cobalto',
       boxInput(p.maxWidth, p.disabled, p.error)
     ),
-    span: cn('px-2 text-obsidiana w-full select-none'),
+    text: cn('px-2 w-full select-none'),
     label: cn('flex flex-col select-none'),
-    error: cn('px-2 text-xs text-chicle'),
-    required: cn('text-chicle ml-1 mb-1'),
+    error: cn('px-2 clr-error-text'),
+    required: cn('clr-error-text ml-1'),
     container: cn('textarea'),
   };